reusable methods and datatypes inside utilities
